Conditional mounts need addresses set in mount()
Conditional mounts are only allowed to reference addresses that were
locally created during the mount() phase. This makes it possible for
an App to know, when a child App/Route is going to be routed to
during navigation, whether any conditional route should be
rendered/shown or destroyed/hidden (using the child App/Route's
address as an indicator).

It would not be possible for a conditional route to know whether a
route or app in some other App has been rendered, or even if its
address has been registered in the case of a deeper hierarchy, so we
enforce local address checks.