expectedAddresses() requires addressesHandlers()
A non-empty array returned from expectedAddresses() requires an array be
returned from addressesHandlers() of the same length, that contains
functions and/or strings representing member functions on the object.

Handler functions match up by position with an object's expected
addresses. This means an object, when sent data using one of its
addresses, will invoke the function at the same array position of
addressesHandlers() as the position of the address that was referenced
in expectedAddresses().

The idea behind this is reflected in this example:

class MyRoute {
    expectedAddresses() {
        ['userInfo', 'documentInfo'];
    }
    addressesHandlers() {
        ['handleUserInfo', function handleDocumentInfo(data){}];
    }
    handleUserInfo(data) {
        // handle data
    }
}
class MyApp {
    mount() {
        return {
            'abc': MyRoute.addresses('userInfo', 'documentInfo'),
        };
    }
}

When `this.sendTo('userInfo', data);` is called from anywhere in the
entire application, The `abc` MyRoute instance will be found, and
its `handleUserInfo` member fn will be invoked with `data` as an arg.

When `this.sendTo('documentInfo', data);` is called from anywhere in the
entire application, The `abc` MyRoute instance will be found, and
the function `handleDocumentInfo` inside the array will be invoked with
`data` as an arg.

Note: The addressesHandlers() function may be called multiple times,
      so no functions defined within should have side-effects.