Addressable Modifier no longer modifies args
MountMapper classes handle all modifier-type scenarios. This is needed
because there are some actions that need to take place depending on the
modifier, and duplicating modifier data to the modifier member data and
opts it passes on create is less intuitive than handling all modifier
situations in one place: the MountMapper, where modifier#create() is
actually called.