MountMapper classes only send expected params
If an App has more params than a child App or Route expects, only the
params that the child expects are passed. This allows a child App or
Route to be mounted deeper in the navigation tree without having to list
ALL params accumulated up to that point. In effect, this makes mouting
an app from a third-party easier, which is what this framework is all
about.

In other words, a child App or Route that only cares about certain
params being available should not error on construction if more params
are available, and should only be sent what it expects to get from the
available list of params.

Note that an App or Route not receiving all expected params will still
throw an error on construction.