diffObjects: Allow undefined if !obj.hasOwn(prop)
When doing a diff on e.g. querystring parameters, some params may be
removed/added and will cause an object diff to return `undefined` for
those properties because they don't exist on the previous or
currently-parsed query params object. In those cases it's necessary to
inform the user that the property has vanished or that it was added.
`undefined` is uniquely suited for this purpose because we otherwise
only allow numbers or strings as values.