Bring back Transition class in a limited capacity
I want to be able to use it for canceling a transition and navigating
anew if e.g. links are clicked rapidly, but haven't yet found a way to
elegantly handle the various prerender/deactivate/render promises while
keeping the Ether app in a predictable state that's easy to understand.

So right now I'm just using Transition to be able to queue navigate()
calls.