Allow generating link hrefs from addresses
Routes now have a `linkTo()` method that takes an address, a `params`
object, and an options object and creates a link ready to be plugged
into an HTML anchor href or a call to `RootApp/Route#navigate()`.

The options object allows the user to supply a transformer to transform
how the `params` object's properties map to the params used in
breadcrumbs mounted on Apps. The options object also allows the user to
choose whether the `basePath` supplied to the RootApp constructor is
prepended to the resulting href (useful for actual HTML links) or not
(useful in direct calls to `navigate()`).