Catch Route render cycle promises;make nav succeed
Any errors thrown or returned promises rejected within Route member
fns `prerender`, `render`, and `deactivate` will be caught and the
navigation to or from that route will subsequently succeed. This means
that ensuring a promise resolves and errors are not thrown is the user's
responsibility. We do however let the user know when these things have
happened if they have debug mode turned on.