interceptLinks catches bubbling click events
If e.g. an image wrapped in an anchor is clicked, the click occurs on
the image, not the anchor. Here we search through the parentNodes of the
event target until we find an anchor, and if so, process it.

Also, seems to be *another* thing CasperJS can't test for.. :(